Int32_t type casting and float32_T type casting done.

+++ Application/Binary/DBinaryParser.cpp
@@ -472,7 +472,7 @@ sError_t DBinaryParser::getFloatFromBuffer(uint8_t *buffer, float *value)
 
     error.value = 0u;
 
-    uFloatVal.floatValue = (float)(0);
+    uFloatVal.floatValue = 0.0f;
 
     uFloatVal.byteValue[0] = (uint8_t)(buffer[0]);
     uFloatVal.byteValue[1] = (uint8_t)(buffer[1]);
@@ -520,7 +520,7 @@ sError_t DBinaryParser::getInt32FromBuffer(uint8_t *buffer, int32_t *value)
     uSint32_t uIntVal;
 
     error.value = 0u;
-    uIntVal.int32Value = (int32_t)(0);
+    uIntVal.int32Value = 0;
 
     uIntVal.byteValue[0] = (uint8_t)(buffer[0]);
     uIntVal.byteValue[1] = (uint8_t)(buffer[1]);
@@ -686,7 +686,7 @@ void DBinaryParser::getBufferFromFloat(float *value, uint8_t *buffer)
 {
     uFloat_t uFloatVal;
 
-    uFloatVal.floatValue = (float32_t)(0.0);
+    uFloatVal.floatValue = 0.0f;
     uFloatVal.floatValue = *value;
     buffer[0] = uFloatVal.byteValue[0];
     buffer[1] = uFloatVal.byteValue[1];
